+++ prescient.el
@@ -8,7 +8,7 @@
 ;; Created: 7 Aug 2017
 ;; Package-Requires: ((emacs "25.1"))
 ;; SPDX-License-Identifier: MIT
-;; Version: 4.1
+;; Version: 5.0
 
 ;;; Commentary: